Move WinningHand model to separate app
