+++ rounds/admin.py
@@ -1,9 +1,8 @@
 from django.contrib import admin
-from .models import DrawnRound, Round, HaiPai, WinningHand, Fuuro, RoundResult
+from .models import DrawnRound, Round, HaiPai, Fuuro, RoundResult
 
 admin.site.register(Round)
 admin.site.register(HaiPai)
-admin.site.register(WinningHand)
 admin.site.register(DrawnRound)
 admin.site.register(Fuuro)
 admin.site.register(RoundResult)
