+++ rounds/models.py
@@ -29,20 +29,6 @@ class HaiPai(models.Model):
     hai_pai = models.CharField('hai pai',max_length=255,null=True,blank=True)
     syanten = models.IntegerField('syanten',editable=False,null=True,blank=True)
 
-class WinningHand(models.Model):
-    WINNING_TYPE = [
-        ('TSUMO','ツモ'),
-        ('RON','ロン'),
-    ]
-    round = models.ForeignKey(Round,on_delete=models.CASCADE)
-    member = models.ForeignKey(Member,on_delete=models.CASCADE)
-    winning_type = models.CharField('tsumo ron',max_length=255,choices=WINNING_TYPE)
-    is_dealer = models.BooleanField('is dealer')
-    winning_hands = models.CharField('winning hands',max_length=255)
-    fuu = models.IntegerField('fuu')
-    faan = models.IntegerField('faan')
-    winning_score = models.IntegerField('winning score',editable=False)
-
 class DrawnRound(models.Model):
     DRAWN_HANDS_TYPE = [
         ('TENPAI','聴牌'),
